pyproject.toml
[tool.poetry]
name = "comfyui-griptape"
version = "2.0.11"
description = "Griptape LLM(Large Language Model) Nodes for ComfyUI."
authors = ["Jason Schleifer <jason.schleifer@gmail.com>"]
readme = "README.md"

#  Used by Comfy Registry https://comfyregistry.org
[project]
name = "comfyui-griptape"
description = "Griptape LLM(Large Language Model) Nodes for ComfyUI."
version = "2.0.11" 
license = {file = "LICENSE"}
dependencies = ["griptape[all]==^1.0.2", "python-dotenv", "poetry==1.8.4", "griptape-black-forest @ git+https://github.com/griptape-ai/griptape-black-forest.git"]

[project.urls]
Repository = "https://github.com/griptape-ai/ComfyUI-Griptape"


[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
griptape = { version = "^1.0.2", extras = ["all"]}
griptape-black-forest = {git = "https://github.com/griptape-ai/griptape-black-forest.git"}


[tool.poetry.group.dev]
optional = true


[tool.poetry.group.dev.dependencies]
icecream = "^2.1.3"
ruff = "^0.8.0"
pyright = "^1.1.376"
pre-commit = "^4.0.0"
typos = "^1.22.9"

#torch = {version = "^2.3.1+cu121", source = "pytorch-gpu-src"}
#torchvision = {version = "^0.18.1+cu121", source = "pytorch-gpu-src"}
#torchaudio = {version = "^2.3.1+cu121", source = "pytorch-gpu-src"}
comfy-cli = "^1.1.5"

[[tool.poetry.source]]
name = "pytorch-gpu-src"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"


[tool.comfy]
PublisherId = "shhlife"
DisplayName = "ComfyUI-Griptape"
Icon = "https://cdn.prod.website-files.com/65d658559223871198e78bca/65e4fd64fb258c6d9bd8c276_griptape-Chunck_beige.svg"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
